rename: 일반적으로 사용되는 클래스를 더 적절한 경로로 이동
