+++ src/main/java/kr/binarybard/hireo/web/member/domain/Member.java
@@ -1,5 +1,6 @@
 package kr.binarybard.hireo.web.member.domain;
 
+import kr.binarybard.hireo.common.BaseTimeEntity;
 import lombok.Builder;
 import org.springframework.security.crypto.password.PasswordEncoder;
